Lint
--HG--
branch : keith

+++ env/env_msg.c
@@ -95,7 +95,7 @@ __wt_mb_add(WT_MBUF *mbp, const char *fmt, ...)
 
 	va_start(ap, fmt);
 
-	current = mbp->next - mbp->first;
+	current = (size_t)(mbp->next - mbp->first);
 	remain = mbp->len - current;
 	len = 64;
 	for (;;) {
@@ -114,7 +114,7 @@ __wt_mb_add(WT_MBUF *mbp, const char *fmt, ...)
 		 * Format the user's information.  If it doesn't fit into the
 		 * buffer we have, re-allocate enough memory and try again.
 		 */
-		len = vsnprintf(mbp->next, remain, fmt, ap);
+		len = (size_t)vsnprintf(mbp->next, remain, fmt, ap);
 		if (len < remain) {
 			mbp->next += len;
 			break;
